chore: consolidate model utils and drive-by cleanups (#476)
