+++ codex-cli/src/hooks/use-confirmation.ts
@@ -1,4 +1,3 @@
-// use-confirmation.ts
 import type { ReviewDecision } from "../utils/agent/review";
 import type React from "react";
